Added empty default authoriziation
The default authorization fails when we go to connect to the client if the
default authorization is not set. This adds a simple `None` default for the
HTTPHandler's private authorization field.

Also a small bit to help people trying to run unit tests.

Incidently, I did not add a unit test to test for this.